🏗️ Architect: Modularized apts/optics/base.py
Refactored the 580-line `OpticalPath` 'God Object' by extracting its specialized responsibilities into a modular MixIn architecture within `apts/optics/models/`.

Key improvements:
- Created `GeometricMixIn` in `apts/optics/models/geometric.py` for zoom, FOV, and brightness logic.
- Created `ResolutionMixIn` in `apts/optics/models/resolution.py` for pixel scale, sampling, and diffraction limits.
- Created `MechanicalMixIn` in `apts/optics/models/mechanical.py` for mass, backfocus, and thermal drift.
- Extended `PhotometryMixIn` with saturation and dynamic range methods.
- Reduced cyclomatic complexity of `apts/optics/base.py` by delegating logic to focused components.
- Maintained full public API compatibility and verified functional integrity with existing unit tests.
